refactor: 🗄️ Slightly change file structure, all private components now in internal.ts files and test library is now in tests/lib/ directory
