+++ lib/internal.ts
@@ -1,4 +1,7 @@
+import { createSocket, Socket as UdpSocket, type SocketOptions as UdpSocketOptions } from 'node:dgram';
+import { Socket as TcpSocket, type TcpSocketConnectOpts as TcpSocketOptions } from 'node:net';
 import { homepage } from '../package.json';
+import { LookupController, LookupOneCallback } from './lookup-controller';
 
 /**
  * Generate documentation link for given urlPath string.
