+++ lib/lookup-controller/super-lookup-controller/internal.ts
