+++ tests/lib/assert-uuid.ts
