+++ tests/lib/dns.ts
